fix: fixed ssr issue
