+++ src/pages/index.tsx
@@ -1,6 +1,8 @@
-import HomePage from '@pages/HomePage';
 import type { NextPage } from 'next';
+import dynamic from 'next/dynamic';
 import React from 'react';
+
+const HomePage = dynamic(() => import('@pages/HomePage'), { ssr: false });
 
 const Index: NextPage = () => {
 	return (
